refactor: replace double with BigDecimal for price calculations to improve monetary accuracy
- Updated Vehicle model to store price as BigDecimal
- Adjusted all relevant code in DealershipFileManager, SalesContract, LeaseContract, and UserInterface
- Ensured consistent use of BigDecimal for calculations involving monetary values
- Improved precision in tax, fee, and lease/sale payment calculations

+++ src/main/java/contracts/Contract.java
@@ -4,6 +4,8 @@ package contracts;
 import lombok.Getter;
 import models.Vehicle;
 
+import java.math.BigDecimal;
+
 @Getter
 public abstract class Contract {
     private final String date;
@@ -18,6 +20,6 @@ public abstract class Contract {
         this.vehicle = vehicle;
     }
 
-    public abstract double getTotalPrice();
-    public abstract double getMonthlyPayment();
+    public abstract BigDecimal getTotalPrice();
+    public abstract BigDecimal getMonthlyPayment();
 }
